PHP Code for Importing Student Mostly finished but may need to iron out some bugs.
One issue is that entering student ID doesn't have a restriction on how long or short the input is. It should be exactly 9 characters long.

Also want to look into maybe fixing user input for phone number because it is a little tedious at the moment. Users have to manually type the hyphens between the numbers